fix(organization): file organization
