+++ frontend/app/src/components/global-components/chat.tsx
@@ -4,16 +4,16 @@ import { Link, useNavigate } from 'react-router-dom';
 import Navbar from '../section-components/navbar';
 import SideBox from '../section-components/side-box';
 import CenterBox from '../section-components/center-box';
-import ChatBox from '../section-components/chat-box';
-import Message from '../section-components/message';
+import ChatBox from '../section-components/chat/chat-box';
+import Message from '../section-components/chat/message';
 import BackgroundGeneral from '../../img/disco2.png';
 import DropDownButton from '../section-components/drop-down-button';
 import UsersList from '../section-components/users-list';
-import ChannelHeader from '../section-components/channel-header';
+import ChannelHeader from '../section-components/chat/channel-header';
 import { User } from '../global-components/interface';
 import { useEffect } from 'react';
 import useUserInfo from '../query-hooks/useUserInfo';
-import MyChannelsList from '../section-components/my-channels-list';
+import MyChannelsList from '../section-components/chat/my-channels-list';
 
 const chanUsersData: User[] = [
   {
