+++ frontend/app/src/components/global-components/profile.tsx
@@ -6,9 +6,9 @@ import Background from '../section-components/background';
 import SideBox from '../section-components/side-box';
 import { MatchData } from '../global-components/interface';
 import FriendList from '../section-components/friend-list';
-import StatsBox from '../section-components/stats-box';
+import StatsBox from '../section-components/profile/stats-box';
 import { useParams } from 'react-router-dom';
-import ProfileBox from '../section-components/profile-box';
+import ProfileBox from '../section-components/profile/their-profile';
 import { useNavigate } from 'react-router-dom';
 import useUserInfo from '../query-hooks/useUserInfo';
 import { useEffect } from 'react';
